UserPermissions/requirements.txt
autopep8==1.2.4
blessings==1.6
bpython==0.15
curtsies==0.2.6
Django==1.9.9
django-guardian==1.4.5
django-sql-explorer==1.0
greenlet==0.4.10
MySQL-python==1.2.5
pep8==1.7.0
pkg-resources==0.0.0
Pygments==2.1.3
requests==2.11.0
six==1.10.0
sqlparse==0.2.1
unicodecsv==0.14.1
wcwidth==0.1.7
